fix #282246: "other" appearance in time signature properties does not work
This bug happens because right after the numerator and denominator strings are set to the style of the chosen "other" timesig, they're both rewrote as numbers due to a logical flaw (this can work because "other" timesig also has a "NORMAL" TimeSigStyle).

Also, set the denominator string to empty won't work as simple as it is, because if the program finds a string empty it'll replace it with a number (like 4 in default). So this fix changes the layout rule of "other" timesig to the same as 4/4 and alla breve, which guarantees the timesig to be set at the middle of the staff.
